\name{Viridot-package}
\alias{Viridot-package}
\alias{Viridot}
\docType{package}
\title{
\packageTitle{Viridot}
}
\description{
\packageDescription{Viridot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Viridot}
\packageIndices{Viridot}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Viridot}

Maintainer: \packageMaintainer{Viridot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
